KEYCLOAK-14709 removed circular dependency
